Rename discretizations module to models
